refactor: 밀부 클래스 이름 변경 및 패키지 이동
