+++ src/main/kotlin/lotto/generator/ticket/ActualLottoShop.kt
@@ -1,0 +1,8 @@
+package lotto.generator.ticket
+
+import lotto.domain.LottoTickets
+
+object ActualLottoShop : LottoShop {
+    override fun provideAutoTickets(ticketCount: Int): LottoTickets = AutoTicketGenerator.provide(ticketCount)
+    override fun provideManualTickets(ticketCount: Int): LottoTickets = ManualTicketGenerator.provide(ticketCount)
+}
