+++ src/main/kotlin/lotto/generator/ticket/LottoShop.kt
@@ -1,4 +1,4 @@
-package lotto.provider.ticket
+package lotto.generator.ticket
 
 import lotto.domain.LottoTickets
 
@@ -7,6 +7,6 @@ interface LottoShop {
     fun provideManualTickets(ticketCount: Int): LottoTickets
 }
 
-interface TicketProvideStrategy {
+interface TicketGenerator {
     fun provide(ticketCount: Int): LottoTickets
 }
