+++ src/test/kotlin/lotto/LottoAutoTest.kt
@@ -7,8 +7,8 @@ import lotto.domain.LottoTicket
 import lotto.domain.LottoTickets
 import lotto.domain.Rank
 import lotto.domain.WinningNumber
-import lotto.provider.ticket.ExactTicketProvideStrategy
-import lotto.provider.ticket.MockLottoShop
+import lotto.generator.ticket.ExactTicketGenerator
+import lotto.generator.ticket.MockLottoShop
 import lotto.view.MockInputView
 import lotto.view.ResultView
 
@@ -26,7 +26,7 @@ class LottoAutoTest : StringSpec({
             resultView = ResultView(),
         ).simulate(
             lottoShop = MockLottoShop(
-                ExactTicketProvideStrategy(
+                ExactTicketGenerator(
                     LottoTickets(
                         listOf(
                             LottoTicket(listOf(1, 2, 3, 4, 5, 6).map { LottoNumber(it) })
@@ -49,7 +49,7 @@ class LottoAutoTest : StringSpec({
             resultView = ResultView(),
         ).simulate(
             lottoShop = MockLottoShop(
-                ExactTicketProvideStrategy(
+                ExactTicketGenerator(
                     LottoTickets(
                         listOf(
                             LottoTicket(listOf(1, 2, 3, 4, 5, 6).map { LottoNumber(it) })
@@ -72,7 +72,7 @@ class LottoAutoTest : StringSpec({
             resultView = ResultView(),
         ).simulate(
             lottoShop = MockLottoShop(
-                ExactTicketProvideStrategy(
+                ExactTicketGenerator(
                     LottoTickets(
                         listOf(
                             LottoTicket(listOf(1, 2, 3, 4, 5, 19).map { LottoNumber(it) })
