+++ src/test/kotlin/lotto/generator/ticket/ExactTicketGenerator.kt
@@ -1,0 +1,7 @@
+package lotto.generator.ticket
+
+import lotto.domain.LottoTickets
+
+class ExactTicketGenerator(private val lottoTickets: LottoTickets) : TicketGenerator {
+    override fun provide(ticketCount: Int): LottoTickets = lottoTickets
+}
