+++ src/test/kotlin/lotto/generator/ticket/MockLottoShop.kt
@@ -1,0 +1,8 @@
+package lotto.generator.ticket
+
+import lotto.domain.LottoTickets
+
+class MockLottoShop(private val ticketGenerateStrategy: ExactTicketGenerator) : LottoShop {
+    override fun provideAutoTickets(ticketCount: Int): LottoTickets = ticketGenerateStrategy.provide(ticketCount)
+    override fun provideManualTickets(ticketCount: Int): LottoTickets = LottoTickets(listOf())
+}
